% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argos.R
\docType{package}
\name{Argos}
\alias{Argos}
\alias{Argos-package}
\title{Argos: Affordances for Retrieval of General OHDSI Stuff}
\description{
The Argos package provides a simple toolset to make accessing databases
that implement the OMOP/OHDSI CDM, and linking clinical data to the
standard vocabulary, easier.
}

