% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_to_std_concepts.R
\name{codes_to_std_concepts}
\alias{codes_to_std_concepts}
\title{Retrieve standard concept IDs for a set of codes}
\usage{
codes_to_std_concepts(codes, vocabulary = NA, src = ohdsi_default_src())
}
\arguments{
\item{codes}{The source codes to map.  This may be a simple character
vector, in which case the \code{vocabulary} parameter is required, or
a data frame, in which case the \code{concept_code} column provides
the source code and the \code{vocabulary_id} column supplies the vocabulary.}

\item{vocabulary}{A character vector listing the vocabulary(ies) from which
the source codes are drawn}

\item{src}{The \code{\link[dplyr]{src}} through which the connection to the
OHDSI vocabularies is made, if different from \code{ohdsi_default_src()}.}
}
\value{
A \code{\link[dplyr]{tbl}} object containing information about
  the standard concepts, as described.
}
\description{
Given a set of codes drawn from one or more terminologies in the OHDSI
vocabulary, return the corresponding standard concept IDs, names, and
vocabularies.
}
\details{
\emph{N.B.} A source code may map to more than one standard concept.

Because the source and standard terms have similar attributes, the
return value disambiguates them by using different names:
\describe{
  \item{src_concept_code}{The code passed into the function.}
  \item{src_vocabulary_id}{The name of the terminology from which
    to the souce code was drawn, also passed into the function.}
  \item{std_concept_code}{The code for the corresponding standard
    concept in its own terminology (e.g. RxNorm)}
  \item{std_vocabulary_id}{The name of the terminology from which
    the standard concept was drawn.}
  \item{std_concept_name}{The human-readable description of the standard
    term, as defined by the terminology from which it was drawn.}
  \item{std_concept_id}{The OHDSI internal ID for the standard concept;
    this is what you use when looking up the term in OHDSI databases}
}
}
\examples{
\dontrun{
codes_to_std_concepts(c('282.61', '282.62', '282.41', '282.42'),
                      vocabulary = 'ICD9CM')
}
}

