% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_argos.R
\name{read_config_file}
\alias{read_config_file}
\title{Read the content of a JSON config file}
\usage{
read_config_file(paths = NA)
}
\arguments{
\item{paths}{A vector of path specifications to use in searching for
the configuration file.}
}
\value{
A data frame containing the contents of the configuration file,
  as parsed by \code{\link[jsonlite]{fromJSON}}, or \code{NA} if no file
  could be read.
}
\description{
Given a vector of file paths, \code{read_config_file} attempts to read a JSON
configuration file, using each element of the vector to locate the file.
It returns the contents of the first file successfully read; subsequent path
specifications are ignored.
}
\examples{
read_config_file( c('~/my_app.json', '~/my_proj.json'))
}

